% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_number.R
\name{ArrayDerivative}
\alias{ArrayDerivative}
\title{Calculate derivative of an tabular function.}
\usage{
ArrayDerivative(x, y, lag = 1)
}
\arguments{
\item{x}{vector of x values, where the function was calculated.}

\item{y}{vector of function values on x.}

\item{lag}{smoothing parameter for derivative calculation.}
}
\value{
Vector of tabular function derivatives.
}
\description{
Calculate derivative of an tabular function.
}
