% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_number.R
\name{EstimateCellsNumber}
\alias{EstimateCellsNumber}
\title{Get the coordinates of the longest true seqeunce of a logical vector.}
\usage{
EstimateCellsNumber(umi.counts, lag = 0.05)
}
\arguments{
\item{umi.counts}{vector with the number of UMIs per cell.}

\item{lag}{smoothing parameter for derivative calculation.}
}
\value{
List with the estimated number of cells:
  \item{min}{minimal number of cells}
  \item{estimated}{expected number of cells}
  \item{max}{maximal number of cells}
}
\description{
Get the coordinates of the longest true seqeunce of a logical vector.
}
