% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{EstimateSaturation}
\alias{EstimateSaturation}
\title{Estimate library saturation with preseqR}
\usage{
EstimateSaturation(reads.by.umig.vec, reads.by.umig.cbs, umi.counts,
  steps.num = 100, max.estimate.rate = 10, top.cells = 1000)
}
\value{
list
}
\description{
Estimate library saturation with preseqR
}
