% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FillAdjacentUmisData}
\alias{FillAdjacentUmisData}
\title{Fill information about adjacent UMIs, their probabilities and differences for each UMI.}
\usage{
FillAdjacentUmisData(umi_probabilites, adjacent_only = FALSE,
  show_progress = FALSE)
}
\arguments{
\item{umi_probabilites}{vector of UMI probabilities.}

\item{adjacent_only}{logical, return only the list of adjacent UMIs.}

\item{show_progress}{show progress bar.}
}
\value{
List with the information about adjacent UMIs.
}
\description{
Fill information about adjacent UMIs, their probabilities and differences for each UMI.
}
