% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{FractionSmoothScatter}
\alias{FractionSmoothScatter}
\title{Plot smoothScatter of geneset fraction for each cell by cell rank}
\usage{
FractionSmoothScatter(fraction, plot.threshold = F, main = "")
}
\description{
Plot smoothScatter of geneset fraction for each cell by cell rank
}
