% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_number.R
\name{GetLongestTrue}
\alias{GetLongestTrue}
\title{Get the coordinates of the longest true seqeunce of a logical vector.}
\usage{
GetLongestTrue(arr)
}
\arguments{
\item{arr}{a logical vector.}
}
\value{
List with two coordinates:
  \item{start}{start position}
  \item{end}{end position}
}
\description{
Get the coordinates of the longest true seqeunce of a logical vector.
}
