% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lq_cells_data}
\alias{lq_cells_data}
\title{Low-quality cells}
\format{
An object of class \code{list} of length 2.
}
\usage{
lq_cells_data
}
\description{
Low-quality cells
}
\keyword{datasets}
