% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lq_cells_info}
\alias{lq_cells_info}
\title{Results of processing of SRR3879617 dataset, neccesary for low-quality cells filtration.}
\format{A list with the next fields:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{https://trace.ncbi.nlm.nih.gov/Traces/sra/?run=SRR3879617}
}
\usage{
lq_cells_info
}
\description{
Results of processing of SRR3879617 dataset, neccesary for low-quality cells filtration.
}
\keyword{datasets}
