% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_quality_cells.R
\name{Normalize}
\alias{Normalize}
\title{Normalize numeric data.}
\usage{
Normalize(x, func = sd)
}
\arguments{
\item{x}{numeric matrix or data.frame.}

\item{func}{Function, which returns a normalizer for given vector.}
}
\value{
Data frame with the normalized x.
}
\description{
Normalize numeric data.
}
\examples{
Normalize(data.frame(x=c(1, 2, 3), y=c(0, 25, 100)))
Normalize(data.frame(x=c(1, 2, 3), y=c(0, 25, 100)), func=function(v) diff(range(v)))
}
