% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ParseUmisPerGene}
\alias{ParseUmisPerGene}
\title{Parse UMIgs.}
\usage{
ParseUmisPerGene(reads_per_umigs, umi_length)
}
\arguments{
\item{reads_per_umigs}{data from the estimation step.}

\item{umi_length}{length of UMI.}
}
\value{
List of lists of vectors with number of reads per UMI per gene per cell.
}
\description{
Parse UMIgs.
}
