% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_number.R
\name{PlotCellsNumberHist}
\alias{PlotCellsNumberHist}
\title{Plot for the estimation of minimal size of high-quality cells.}
\usage{
PlotCellsNumberHist(umi.counts, breaks = 100, title = NULL,
  estimate.cells.number = F, alpha = 0.6, show.legend = T)
}
\arguments{
\item{umi.counts}{vector with the number of UMIs per cell.}

\item{breaks}{number of breaks in a histogram.}

\item{title}{title of the plot.}

\item{alpha}{alpha parameter of the plot}
}
\value{
Plot of ggplot type.
}
\description{
Plot for the estimation of minimal size of high-quality cells.
}
