% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_number.R
\name{PlotCellsNumberLogLog}
\alias{PlotCellsNumberLogLog}
\title{Plot for the estimation of minimal size of high-quality cells in log10 scale}
\usage{
PlotCellsNumberLogLog(
  umi.counts,
  estimate.cells.number = FALSE,
  show.legend = TRUE,
  gg.base = NULL,
  plot.label = NULL,
  plot.border = TRUE,
  linewidth = 1,
  alpha = 1,
  logticks = TRUE
)
}
\arguments{
\item{umi.counts}{vector with the number of UMIs per cell}

\item{estimate.cells.number}{boolean Whether to calculate cell numbers via EstimateCellsNumber() (default=FALSE)}

\item{plot.border}{boolean Whether to include plot border (default=TRUE)}

\item{linewidth}{numeric Line width in ggplot2, via the argument 'size' in 'geom_line()' (default=1)}

\item{alpha}{numeric Value of the alpha parameter of the plot (default=1.0)}

\item{logticks}{boolean Whether to include log plot tick marks (default=TRUE)}
}
\value{
Plot of ggplot type
}
\description{
Plot for the estimation of minimal size of high-quality cells in log10 scale
}
