% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_number.R
\name{PlotCellsNumberSummary}
\alias{PlotCellsNumberSummary}
\title{Plot for the summary of cell numbers (Deprecated).}
\usage{
PlotCellsNumberSummary(umi.counts, breaks = 100, mask = c(T, T, T))
}
\arguments{
\item{umi.counts}{vector with the number of UMIs per cell.}

\item{breaks}{number of breaks in a histogram.}

\item{mask}{mask for the three plots}
}
\description{
Plot for the summary of cell numbers (Deprecated).
}
