% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{PlotGenesPerCell}
\alias{PlotGenesPerCell}
\title{Plot distribution of log number of genes by cells}
\usage{
PlotGenesPerCell(count.matrix, bins = 50)
}
\arguments{
\item{bins}{number of bins in histogram}
}
\value{
ggplot object with the plot.
}
\description{
Plot distribution of log number of genes by cells
}
