% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{PlotIntergenicFractionByChromosomes}
\alias{PlotIntergenicFractionByChromosomes}
\title{Plot fraction of different read types for each chromosome}
\usage{
PlotIntergenicFractionByChromosomes(reads.per.chr.per.cells,
  chromosome.reads.threshold = 0.001)
}
\arguments{
\item{reads.per.chr.per.cells}{named list, which contains data frames for each type of reads}

\item{chromosome.reads.threshold}{minimal fraction of reads for chromosome to be plotted}
}
\value{
ggplot object with the plot.
}
\description{
Plot fraction of different read types for each chromosome
}
