% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{PlotReadsPerUmiByCells}
\alias{PlotReadsPerUmiByCells}
\title{Plot mean number of reads per UMI for each cell by cell rank}
\usage{
PlotReadsPerUmiByCells(mean.reads.per.umi, umi.counts, ...)
}
\arguments{
\item{...}{additional parameters of smoothScatter}
}
\description{
Plot mean number of reads per UMI for each cell by cell rank
}
