% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{PlotSaturationEstimates}
\alias{PlotSaturationEstimates}
\title{Plot estimated library saturation}
\usage{
PlotSaturationEstimates(preseq.estimates)
}
\arguments{
\item{preseq.estimates}{named list of results of EstimateSaturation calls.}
}
\value{
ggplot object with the plot.
}
\description{
Plot estimated library saturation
}
