% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{PlotUmisDistribution}
\alias{PlotUmisDistribution}
\title{Plot distribution of UMI}
\usage{
PlotUmisDistribution(reads.per.umi.per.cb, trim.quantile = 0.99, bins = 50)
}
\arguments{
\item{trim.quantile}{(default=0.99)}

\item{bins}{number of bins in histogram (default=50)}
}
\value{
ggplot object with the plot of UMI distribution
}
\description{
Plot distribution of UMI
}
