% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_classifier.R
\name{PredictKDE}
\alias{PredictKDE}
\title{Predict data with the KDE classifier.}
\usage{
PredictKDE(clf, x, bandwidth.mult = 1)
}
\arguments{
\item{clf}{trained classifier.}

\item{x}{data for classification.}

\item{bandwidth.mult}{multiplier for KDE bandwidth}
}
\value{
Class probabilities.
}
\description{
Predict data with the KDE classifier.
}
