% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_number.R
\name{PrepareCellsNumberPlot}
\alias{PrepareCellsNumberPlot}
\title{Prepare data for plotting the number of cells}
\usage{
PrepareCellsNumberPlot(
  umi.counts,
  breaks,
  estimate.cells.number = FALSE,
  title = NULL
)
}
\arguments{
\item{umi.counts}{vector with the number of UMIs per cell}

\item{breaks}{numeric Number of breaks in a histogram}

\item{estimate.cells.number}{boolean Whether to calculate cell numbers via EstimateCellsNumber() (default=FALSE)}

\item{title}{title of the plot (default=NULL)}
}
\value{
List with the data, needed for plotting:
  \item{plot.df}{data frame with the data for plotting}
  \item{title}{title of th plot}
  \item{umi.counts}{transformed number of UMIs per cell}
  \item{y.label}{y label of the plot}
}
\description{
Prepare data for plotting the number of cells
}
