% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reads_per_umi_per_cb}
\alias{reads_per_umi_per_cb}
\title{Results of processing of 10x Frozen BMMCs (Healthy Control 1) dataset, neccesary for UMI errors correction.}
\format{An object of class \code{list} of length 144013.}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets/1.1.0/frozen_bmmc_healthy_donor1}
}
\usage{
reads_per_umi_per_cb
}
\description{
Results of processing of 10x Frozen BMMCs (Healthy Control 1) dataset, neccesary for UMI errors correction.
}
\keyword{datasets}
