% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reads_per_umi_per_cell}
\alias{reads_per_umi_per_cell}
\title{Results of processing of mouse BC dataset (SCG71), neccesary for UMI errors correction for the 100 largest cells.}
\format{An object of class \code{list} of length 100.}
\usage{
reads_per_umi_per_cell
}
\description{
Results of processing of mouse BC dataset (SCG71), neccesary for UMI errors correction for the 100 largest cells.
}
\keyword{datasets}
