% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_classifier.R
\name{ScoreCells}
\alias{ScoreCells}
\title{Score cells with a KDE classifier}
\usage{
ScoreCells(data, cells.quality)
}
\arguments{
\item{data}{data for classification}

\item{cells.quality}{vector of cell barcodes, named with cells.quality ("High", "Low")}
}
\value{
Probability of cell to be high-quality
}
\description{
Score cells with a KDE classifier
}
