% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_quality_cells.R
\name{ScorePipelineCells}
\alias{ScorePipelineCells}
\title{Score cells with a KDE classifier.}
\usage{
ScorePipelineCells(pipeline.data, filter.high.mit.fraction = F,
  mitochondrion.genes = NULL, mit.chromosome.name = NULL,
  mit.threshold = NULL, tags.data = NULL, predict.all = FALSE,
  r2.threshold = 0.05, pcs.number = 3)
}
\arguments{
\item{pipeline.data}{data for classification.}
}
\value{
Probability of cell to be high-quality.
}
\description{
Score cells with a KDE classifier.
}
