% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_classifier.R
\name{TrainClassifier}
\alias{TrainClassifier}
\title{Train classifier with TrainKDE}
\usage{
TrainClassifier(
  tr.data,
  cells.quality,
  umi.counts = NULL,
  trim.low.quality.rate = 1.5,
  prior.probs = c(0.5, 0.5)
)
}
\arguments{
\item{tr.data}{training data for classification}

\item{cells.quality}{vector of cell barcodes, named with cells.quality ("High", "Low")}

\item{umi.counts}{vector with the number of UMIs per cell (default=NULL)}

\item{trim.low.quality.rate}{numeric Value at which to filter low quality values (default=1.5)}

\item{prior.probs}{Prior probabilities or weights of classes (default=c(0.5, 0.5))}
}
\value{
Trained KDE classifier
}
\description{
Train classifier with TrainKDE
}
