% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_classifier.R
\name{TrainKDE}
\alias{TrainKDE}
\title{Train binary KDE classifier.}
\usage{
TrainKDE(x, y, H = ks::Hns, prior.probs = c(0.5, 0.5))
}
\arguments{
\item{x}{train data.}

\item{y}{train answers.}

\item{H}{the algorithm of bandwidth selection (see 'ks' documentation).}

\item{prior.probs}{Prior probabilities or weights of classes.}
}
\value{
Trained KDE classifier.
}
\description{
Train binary KDE classifier.
}
