% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_quality_cells.R
\name{GetOptimalPcs}
\alias{GetOptimalPcs}
\title{Perform PCA with the optimal number of principal components.}
\usage{
GetOptimalPcs(
  data,
  explained.var.required = 0.98,
  max.pcs = 3,
  loadings.filt.threshold = 0.075
)
}
\arguments{
\item{data}{data for low-quality cells filtration}

\item{explained.var.required}{numeric Minimal fraction of explained variance (default=0.98)}

\item{max.pcs}{numeric Maximal number of output principal components (default=3)}

\item{loadings.filt.threshold}{numeric Minimal contribution to loadings, with which feature is considered as used (default=7.5e-2)}
}
\value{
List with the PCA results:
  \item{total.variance.explained}{fraction of the explained variance.}
  \item{pca.data}{transformed data with the optimal number of principal components.}
  \item{used.features}{features, which contribute to pca.data.}
}
\description{
Perform PCA with the optimal number of principal components.
}
