% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetUmisDistribution}
\alias{GetUmisDistribution}
\title{Estimate a distribution of observed UMI probabilities.}
\usage{
GetUmisDistribution(umis_per_gene_per_cell, smooth = 1L)
}
\arguments{
\item{umis_per_gene_per_cell}{list of vectors: number of UMIs per gene per cell (zeros can be omitted).}

\item{smooth}{smooth term, which is added to each UMI probability in case if some UMIs have only few observations.}

\item{umi_length}{length of UMI.}
}
\value{
Vector of UMI probabilities.
}
\description{
Estimate a distribution of observed UMI probabilities.
}
