% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{PlotCellScores}
\alias{PlotCellScores}
\title{Plot smoothScatter with cell quality scores by cell rank}
\usage{
PlotCellScores(
  scores,
  cells.number = NULL,
  y.threshold = NULL,
  main = NULL,
  bandwidth = c(length(scores)/100, 0.008)
)
}
\arguments{
\item{cells.number}{(default=NULL)}

\item{y.threshold}{(default=NULL)}

\item{main}{(default=NULL)}

\item{bandwidth}{(default=c(length(scores) / 100, 0.008)))}
}
\value{
smoothScatter plot with cell quality scores by cell rank
}
\description{
Plot smoothScatter with cell quality scores by cell rank
}
