% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_number.R
\name{PlotCellsNumberHist}
\alias{PlotCellsNumberHist}
\title{Plot for the estimation of minimal size of high-quality cells.}
\usage{
PlotCellsNumberHist(
  umi.counts,
  breaks = 100,
  title = NULL,
  estimate.cells.number = FALSE,
  alpha = 0.6,
  show.legend = TRUE
)
}
\arguments{
\item{umi.counts}{vector with the number of UMIs per cell}

\item{breaks}{numeric Number of breaks in a histogram}

\item{title}{title of the plot (default=NULL)}

\item{estimate.cells.number}{boolean Whether to calculate cell numbers via EstimateCellsNumber() (default=FALSE)}

\item{alpha}{numeric Value of the alpha parameter of the plot (default=0.6)}
}
\value{
Plot of ggplot type.
}
\description{
Plot for the estimation of minimal size of high-quality cells.
}
