% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_number.R
\name{PlotCellsNumberLine}
\alias{PlotCellsNumberLine}
\title{Plot for the estimation of the number of high-quality cells.}
\usage{
PlotCellsNumberLine(umi.counts, breaks = 100, title = NULL,
  estimate.cells.number = F, show.legend = T, gg.base = NULL,
  plot.label = NULL)
}
\arguments{
\item{umi.counts}{vector with the number of UMIs per cell.}

\item{breaks}{number of breaks in a histogram.}

\item{title}{title of the plot.}
}
\value{
Plot of ggplot type.
}
\description{
Plot for the estimation of the number of high-quality cells.
}
