% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_number.R
\name{PlotCellsNumberSummary}
\alias{PlotCellsNumberSummary}
\title{Plot for the summary of cell numbers (Deprecated).}
\usage{
PlotCellsNumberSummary(umi.counts, breaks = 100, mask = c(TRUE, TRUE, TRUE))
}
\arguments{
\item{umi.counts}{vector with the number of UMIs per cell}

\item{breaks}{numeric Number of breaks in a histogram}

\item{mask}{mask for the three plots (default=c(TRUE, TRUE, TRUE))}
}
\description{
Plot for the summary of cell numbers (Deprecated).
}
