% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{PlotReadsPerUmiByCells}
\alias{PlotReadsPerUmiByCells}
\title{Plot mean number of reads per UMI for each cell by cell rank}
\usage{
PlotReadsPerUmiByCells(mean.reads.per.umi, umi.counts, ...)
}
\arguments{
\item{...}{additional parameters for smoothScatter()}
}
\value{
smoothScatter plot with the mean number of reads per UMI for each cell by cell rank
}
\description{
Plot mean number of reads per UMI for each cell by cell rank
}
