% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{PlotUmisDistribution}
\alias{PlotUmisDistribution}
\title{Plot distribution of UMI}
\usage{
PlotUmisDistribution(reads.per.umi.per.cb, trim.quantile = 0.99,
  bins = 50)
}
\arguments{
\item{bins}{number of bins in histogram}
}
\value{
ggplot object with the plot.
}
\description{
Plot distribution of UMI
}
