% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_quality_cells.R
\name{PrepareLqCellsData}
\alias{PrepareLqCellsData}
\alias{PrepareLqCellsDataPipeline}
\title{Prepare data frame for low-quality cells filtration.}
\usage{
PrepareLqCellsData(count.matrix, aligned.reads.per.cell,
  total.umis.per.cell = NULL, total.reads.per.cell = NULL,
  intergenic.reads.per.cell = NULL, mitochondrion.fraction = NULL)

PrepareLqCellsDataPipeline(data, total.reads.per.cell = NULL,
  mitochondrion.genes = NULL, mit.chromosome.name = NULL)
}
\arguments{
\item{count.matrix}{Matrix with counts of umi per gene per cell.}

\item{intergenic.reads.per.cell}{Number of intergenic reads per cell.}

\item{mitochondrion.fraction}{Fraction of mitochondrion reads or UMIs per cell}

\item{data}{data, extracted with the pipeline.}

\item{reads.per.gene}{List, which contains number of reads per gene per cell.}

\item{total.reads.per.cb}{Number of reads per cell before alignment.}

\item{merge.targets}{Targets of CB merge. Used only with total.reads.per.cb provided.}

\item{pcs.number}{maximal number of principal components, which is returned. Set to NULL to return raw data.}

\item{merge.targets}{Targets of CB merge. Used only with total.reads.per.cell provided.}
}
\value{
Data frame for low-quality cells filtration.
}
\description{
Prepare data frame for low-quality cells filtration.
The more parameters provided, the more detailed matrix returned.
}
\section{Functions}{
\itemize{
\item \code{PrepareLqCellsDataPipeline}: wrapper for the data, obtained with the dropEst pipeline.
}}

