% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_quality_cells.R
\name{PrepareLqCellsData}
\alias{PrepareLqCellsData}
\alias{PrepareLqCellsDataPipeline}
\title{Prepare data.frame for low-quality cells filtration. The more parameters provided, the more detailed matrix returned.}
\usage{
PrepareLqCellsData(
  count.matrix,
  aligned.reads.per.cell,
  total.umis.per.cell = NULL,
  total.reads.per.cell = NULL,
  intergenic.reads.per.cell = NULL,
  mitochondrion.fraction = NULL,
  scale = TRUE
)

PrepareLqCellsDataPipeline(
  data,
  total.reads.per.cell = NULL,
  merge.targets = NULL,
  mitochondrion.genes = NULL,
  mit.chromosome.name = NULL,
  scale = TRUE
)
}
\arguments{
\item{count.matrix}{Matrix with counts of umi per gene per cell}

\item{aligned.reads.per.cell}{}

\item{total.umis.per.cell}{(default=NULL)}

\item{total.reads.per.cell}{Number of reads per cell before alignment (default=NULL)}

\item{intergenic.reads.per.cell}{Number of intergenic reads per cell (default=NULL)}

\item{mitochondrion.fraction}{Fraction of mitochondrion reads or UMIs per cell (default=NULL)}

\item{scale}{(default=TRUE)}

\item{data}{data, extracted with the pipeline.}

\item{merge.targets}{Targets of CB merge. Used only with total.reads.per.cell provided.}

\item{mitochondrion.genes}{(default=NULL)}
}
\value{
Data.frame for low-quality cells filtration
}
\description{
Prepare data.frame for low-quality cells filtration. The more parameters provided, the more detailed matrix returned.
}
\section{Functions}{
\itemize{
\item \code{PrepareLqCellsDataPipeline}: wrapper for the data, obtained with the dropEst pipeline.
}}

