% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saturation_srr1784312}
\alias{saturation_srr1784312}
\title{Results of saturation analysis of SRR1784312 dataset (mouse ES cells).}
\format{A list with the next fields:
\describe{
  \item{sat}{saturation curve}
  \item{cur}{observed values}
  ...
}}
\source{
\url{https://trace.ncbi.nlm.nih.gov/Traces/sra/?run=SRR1784312}
}
\usage{
saturation_srr1784312
}
\description{
Results of saturation analysis of SRR1784312 dataset (mouse ES cells).
}
\keyword{datasets}
