% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_classifier.R
\name{ScoreCells}
\alias{ScoreCells}
\title{Score cells with a KDE classifier.}
\usage{
ScoreCells(data, cells.quality)
}
\arguments{
\item{data}{data for classification.}

\item{hq.cbs}{vector of cell barcodes, which mostly have high quality.}

\item{lq.cbs}{vector of cell barcodes, which mostly have low quality.}
}
\value{
Probability of cell to be high-quality.
}
\description{
Score cells with a KDE classifier.
}
