% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_quality_cells.R
\name{ScorePipelineCells}
\alias{ScorePipelineCells}
\title{Score cells with a KDE classifier.}
\usage{
ScorePipelineCells(pipeline.data, mitochondrion.genes = NULL,
  mit.chromosome.name = NULL, tags.data = NULL,
  filter.mitochondrial = NULL, filter.intergenic = T,
  mit.fraction.threshold = NULL, intergenic.fraction.threshold = NULL,
  max.pcs.number = 3, predict.all = FALSE, verbose = FALSE,
  kde.bandwidth.mult = 1, cell.number = NULL)
}
\arguments{
\item{pipeline.data}{data for classification.}
}
\value{
Probability of cell to be high-quality.
}
\description{
Score cells with a KDE classifier.
}
